[refactor] member-common 제거 및 application layer port구조 변경
1. member-common 제거
  - 공통 Exception 사용 하지 않도록 변경 및 기본 예외 사용
2. application layer 구조 변경
  - application layer service & port 구조로 사용하기로 협의 그에 따른 변경
