+++ member-api/member-application/src/main/java/org/flab/deliveryplatform/member/application/service/DuplicatedEmailException.java
@@ -1,0 +1,24 @@
+package org.flab.deliveryplatform.member.application.service;
+
+public class DuplicatedEmailException extends RuntimeException {
+
+    public DuplicatedEmailException() {
+    }
+
+    public DuplicatedEmailException(String message) {
+        super(message);
+    }
+
+    public DuplicatedEmailException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public DuplicatedEmailException(Throwable cause) {
+        super(cause);
+    }
+
+    public DuplicatedEmailException(String message, Throwable cause, boolean enableSuppression,
+        boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}
