+++ member-api/member-application/src/main/java/org/flab/deliveryplatform/member/application/service/MemberService.java
@@ -1,14 +1,10 @@
 package org.flab.deliveryplatform.member.application.service;
 
-import static org.flab.deliveryplatform.member.common.exception.MemberErrorCode.DUPLICATED_EMAIL;
-import static org.flab.deliveryplatform.member.common.exception.MemberErrorCode.INVALID_MEMBER_INFO;
-
 import lombok.RequiredArgsConstructor;
-import org.flab.deliveryplatform.common.exception.DeliveryPlatformException;
-import org.flab.deliveryplatform.member.application.persistence.MemberPersistencePort;
-import org.flab.deliveryplatform.member.application.usecase.MemberInfoResult;
-import org.flab.deliveryplatform.member.application.usecase.SignUpMemberCommand;
-import org.flab.deliveryplatform.member.application.usecase.SignUpMemberResult;
+import org.flab.deliveryplatform.member.application.port.MemberPersistencePort;
+import org.flab.deliveryplatform.member.application.port.dto.MemberInfoResult;
+import org.flab.deliveryplatform.member.application.port.dto.SignUpMemberCommand;
+import org.flab.deliveryplatform.member.application.port.dto.SignUpMemberResult;
 import org.flab.deliveryplatform.member.domain.Member;
 import org.springframework.stereotype.Service;
 
@@ -34,21 +30,21 @@ public class MemberService {
 
     private void validateSignUp(SignUpMemberCommand signUpMemberCommand) {
         if (memberPersistencePort.exists(signUpMemberCommand.getEmail())) {
-            throw new DeliveryPlatformException(DUPLICATED_EMAIL);
+            throw new DuplicatedEmailException("중복된 이메일 입니다.");
         }
     }
 
     public MemberInfoResult info(Long memberId) {
         return MemberInfoResult.from(
             memberPersistencePort.findById(memberId).orElseThrow(
-                () -> new DeliveryPlatformException(INVALID_MEMBER_INFO))
+                () -> new IllegalArgumentException("잘못된 회원 정보입니다."))
         );
     }
 
     public MemberInfoResult findByEmailAndPassword(String email, String password) {
         return MemberInfoResult.from(
             memberPersistencePort.findByEmailAndPassword(email, password).orElseThrow(
-                () -> new DeliveryPlatformException(INVALID_MEMBER_INFO))
+                () -> new IllegalArgumentException("잘못된 회원 정보입니다."))
         );
     }
