--- member-api/member-common/build.gradle
@@ -1,5 +1,0 @@
-version '1.0-SNAPSHOT'
-
-dependencies {
-    implementation project(':common')
-}
