+++ member-api/member-infrastructure/src/main/java/org/flab/deliveryplatform/member/infrastructure/persistence/MemberPersistenceAdaptor.java
@@ -2,7 +2,7 @@ package org.flab.deliveryplatform.member.infrastructure.persistence;
 
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.flab.deliveryplatform.member.application.persistence.MemberPersistencePort;
+import org.flab.deliveryplatform.member.application.port.MemberPersistencePort;
 import org.flab.deliveryplatform.member.domain.Member;
 import org.springframework.stereotype.Component;
